move list out of pages
